requirements.txt
tensorflow==2.12.0
keras==2.12.0
numpy==1.23.5
pandas==1.5.3
Pillow==9.5.0
torch==2.0.0
tqdm==4.65.0
torchvision==0.15.1
scikit-learn==1.2.2
opencv-python==4.5.3.56
urllib3==1.26.15
